build: Add disable_vagrant variable to build on the host.

+++ Makefile
@@ -74,6 +74,7 @@ help:
 	@echo "               Current build name: $(_build_name)"
 	@echo "  unstripped - Set to non-empty to avoid stripping installed library files."
 	@echo "  enable_ccache - Enabled by default, set to 0 to disable ccache."
+	@echo "  disable_vagrant - Set to 1 to disable vagrant and build directly on the host."
 	@echo "  protonsdk_version - Version of the proton sdk image to use for building,"
 	@echo "                      use protonsdk_version=local to build it locally."
 	@echo ""
@@ -103,12 +104,21 @@ help:
 	@echo ""
 	@echo "Running out of disk space in the VM? See resize-vagrant-disk.sh"
 
+ifeq ($(disable_vagrant),1)
+VAGRANT_SHELL := $(SHELL)
+VAGRANT_DIR := vagrant_share
+BUILD_DIR := ../$(BUILD_DIR)
+else
 VAGRANT_SHELL := vagrant ssh
+VAGRANT_DIR := /vagrant
+endif
 
 vagrant: private SHELL := $(SHELL)
 vagrant:
+ifneq ($(disable_vagrant),1)
 	vagrant up
 	vagrant rsync debian10
+endif
 
 clean: private SHELL := $(VAGRANT_SHELL)
 clean: vagrant
@@ -131,36 +141,43 @@ endif
 
 downloads: private SHELL := $(VAGRANT_SHELL)
 downloads: configure
+ifneq ($(disable_vagrant),1)
 	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR) downloads
+endif
 
 proton: private SHELL := $(VAGRANT_SHELL)
 proton: downloads
 	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR)/ $(UNSTRIPPED) $(CCACHE_FLAG) dist && \
 	echo "Proton built in VM. Use 'install' or 'deploy' targets to retrieve the build."
 
+ifneq ($(disable_vagrant),1)
+install-internal: private STEAM_DIR := $(VAGRANT_DIR)/
+endif
 install-internal: | vagrant_share/compatibilitytools.d/$(_build_name)
 install-internal: private SHELL := $(VAGRANT_SHELL)
 install-internal: downloads
-	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR)/ $(UNSTRIPPED) $(CCACHE_FLAG) STEAM_DIR=/vagrant/ install
+	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR)/ $(UNSTRIPPED) $(CCACHE_FLAG) STEAM_DIR=$(STEAM_DIR) install
 
 install: install-internal
+ifneq ($(disable_vagrant),1)
 	mkdir -p $(STEAM_DIR)/compatibilitytools.d/
 	rm -rf $(STEAM_DIR)/compatibilitytools.d/$(_build_name)/files/ #remove proton's internal files, but preserve user_settings etc from top-level
 	cp -Rf --no-dereference --preserve=mode,links vagrant_share/compatibilitytools.d/$(_build_name) $(STEAM_DIR)/compatibilitytools.d/
+endif
 	echo "Proton installed to your local Steam installation"
 
 redist: | vagrant_share/$(DEPLOY_DIR)
 redist: private SHELL := $(VAGRANT_SHELL)
 redist: downloads
-	rm -rf /vagrant/$(DEPLOY_DIR)/* && \
-	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR)/ $(UNSTRIPPED) $(CCACHE_FLAG) redist && cp -Rf $(BUILD_DIR)/redist/* /vagrant/$(DEPLOY_DIR) && \
+	rm -rf $(VAGRANT_DIR)/$(DEPLOY_DIR)/* && \
+	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR)/ $(UNSTRIPPED) $(CCACHE_FLAG) redist && cp -Rf $(BUILD_DIR)/redist/* $(VAGRANT_DIR)/$(DEPLOY_DIR) && \
 	echo "Proton build available at vagrant_share/$(DEPLOY_DIR)"
 
 deploy: | vagrant_share/$(DEPLOY_DIR)-deploy
 deploy: private SHELL := $(VAGRANT_SHELL)
 deploy: downloads
-	rm -rf /vagrant/$(DEPLOY_DIR)-deploy/* && \
-	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR)/ $(UNSTRIPPED) $(CCACHE_FLAG) deploy && cp -Rf $(BUILD_DIR)/deploy/* /vagrant/$(DEPLOY_DIR)-deploy && \
+	rm -rf $(VAGRANT_DIR)/$(DEPLOY_DIR)-deploy/* && \
+	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR)/ $(UNSTRIPPED) $(CCACHE_FLAG) deploy && cp -Rf $(BUILD_DIR)/deploy/* $(VAGRANT_DIR)/$(DEPLOY_DIR)-deploy && \
 	echo "Proton deployed to vagrant_share/$(DEPLOY_DIR)-deploy"
 
 module: | vagrant_share/$(module)/lib/wine/
@@ -168,53 +185,53 @@ module: | vagrant_share/$(module)/lib64/wine/
 module: private SHELL := $(VAGRANT_SHELL)
 module: downloads
 	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR)/ $(UNSTRIPPED) $(CCACHE_FLAG) module=$(module) module && \
-	cp -f $(BUILD_DIR)/obj-wine32/dlls/$(module)/$(module)$(MODULE_SFX)* /vagrant/$(module)/lib/wine/ && \
-	cp -f $(BUILD_DIR)/obj-wine64/dlls/$(module)/$(module)$(MODULE_SFX)* /vagrant/$(module)/lib64/wine/ && \
-	find /vagrant/$(module)/ -type f -name "*.dll" -printf "%p\0" | xargs --verbose -0 -r -P8 -n3 proton/make/pefixup.py && \
-	cp -f $(BUILD_DIR)/obj-wine32/dlls/$(module)/$(module).so /vagrant/$(module)/lib/wine/ && \
-	cp -f $(BUILD_DIR)/obj-wine64/dlls/$(module)/$(module).so /vagrant/$(module)/lib64/wine/ && \
-	rm -f /vagrant/$(module)/lib*/wine/*.fake
+	cp -f $(BUILD_DIR)/obj-wine32/dlls/$(module)/$(module)$(MODULE_SFX)* $(VAGRANT_DIR)/$(module)/lib/wine/ && \
+	cp -f $(BUILD_DIR)/obj-wine64/dlls/$(module)/$(module)$(MODULE_SFX)* $(VAGRANT_DIR)/$(module)/lib64/wine/ && \
+	find $(VAGRANT_DIR)/$(module)/ -type f -name "*.dll" -printf "%p\0" | xargs --verbose -0 -r -P8 -n3 proton/make/pefixup.py && \
+	cp -f $(BUILD_DIR)/obj-wine32/dlls/$(module)/$(module).so $(VAGRANT_DIR)/$(module)/lib/wine/ && \
+	cp -f $(BUILD_DIR)/obj-wine64/dlls/$(module)/$(module).so $(VAGRANT_DIR)/$(module)/lib64/wine/ && \
+	rm -f $(VAGRANT_DIR)/$(module)/lib*/wine/*.fake
 
 dxvk: | vagrant_share/dxvk/lib/wine/dxvk
 dxvk: | vagrant_share/dxvk/lib64/wine/dxvk
 dxvk: private SHELL := $(VAGRANT_SHELL)
 dxvk: downloads
 	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR)/ $(UNSTRIPPED) $(CCACHE_FLAG) dxvk && \
-	cp -f $(BUILD_DIR)/dist/files/lib/wine/dxvk/*.dll /vagrant/dxvk/lib/wine/dxvk/ && \
-	cp -f $(BUILD_DIR)/dist/files/lib64/wine/dxvk/*.dll /vagrant/dxvk/lib64/wine/dxvk/ && \
-	find /vagrant/dxvk/ -type f -name "*.dll" -printf "%p\0" | xargs --verbose -0 -r -P8 -n3 proton/make/pefixup.py'
+	cp -f $(BUILD_DIR)/dist/files/lib/wine/dxvk/*.dll $(VAGRANT_DIR)/dxvk/lib/wine/dxvk/ && \
+	cp -f $(BUILD_DIR)/dist/files/lib64/wine/dxvk/*.dll $(VAGRANT_DIR)/dxvk/lib64/wine/dxvk/ && \
+	find $(VAGRANT_DIR)/dxvk/ -type f -name "*.dll" -printf "%p\0" | xargs --verbose -0 -r -P8 -n3 proton/make/pefixup.py'
 
 vkd3d-proton: | vagrant_share/vkd3d-proton/lib/wine/vkd3d-proton
 vkd3d-proton: | vagrant_share/vkd3d-proton/lib64/wine/vkd3d-proton
 vkd3d-proton: private SHELL := $(VAGRANT_SHELL)
 vkd3d-proton: downloads
 	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR)/ $(UNSTRIPPED) $(CCACHE_FLAG) vkd3d-proton && \
-	cp -f $(BUILD_DIR)/dist/files/lib/wine/vkd3d-proton/*.dll /vagrant/vkd3d-proton/lib/wine/vkd3d-proton/ && \
-	cp -f $(BUILD_DIR)/dist/files/lib64/wine/vkd3d-proton/*.dll /vagrant/vkd3d-proton/lib64/wine/vkd3d-proton/ && \
-	find /vagrant/vkd3d-proton/ -type f -name "*.dll" -printf "%p\0" | xargs --verbose -0 -r -P8 -n3 proton/make/pefixup.py'
+	cp -f $(BUILD_DIR)/dist/files/lib/wine/vkd3d-proton/*.dll $(VAGRANT_DIR)/vkd3d-proton/lib/wine/vkd3d-proton/ && \
+	cp -f $(BUILD_DIR)/dist/files/lib64/wine/vkd3d-proton/*.dll $(VAGRANT_DIR)/vkd3d-proton/lib64/wine/vkd3d-proton/ && \
+	find $(VAGRANT_DIR)/vkd3d-proton/ -type f -name "*.dll" -printf "%p\0" | xargs --verbose -0 -r -P8 -n3 proton/make/pefixup.py'
 
 lsteamclient: | vagrant_share/lsteamclient/lib/wine
 lsteamclient: | vagrant_share/lsteamclient/lib64/wine
 lsteamclient: private SHELL := $(VAGRANT_SHELL)
 lsteamclient: downloads
 	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR)/ $(UNSTRIPPED) $(CCACHE_FLAG) lsteamclient && \
-	cp -f $(BUILD_DIR)/dist/files/lib/wine/lsteamclient.dll.so /vagrant/lsteamclient/lib/wine && \
-	cp -f $(BUILD_DIR)/dist/files/lib64/wine/lsteamclient.dll.so /vagrant/lsteamclient/lib64/wine
+	cp -f $(BUILD_DIR)/dist/files/lib/wine/lsteamclient.dll.so $(VAGRANT_DIR)/lsteamclient/lib/wine && \
+	cp -f $(BUILD_DIR)/dist/files/lib64/wine/lsteamclient.dll.so $(VAGRANT_DIR)/lsteamclient/lib64/wine
 
 vrclient: | vagrant_share/vrclient/lib/wine
 vrclient: | vagrant_share/vrclient/lib64/wine
 vrclient: private SHELL := $(VAGRANT_SHELL)
 vrclient: downloads
 	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR)/ $(UNSTRIPPED) $(CCACHE_FLAG) vrclient && \
-	cp -f $(BUILD_DIR)/dist/files/lib/wine/vrclient.dll.so /vagrant/vrclient/lib/wine && \
-	cp -f $(BUILD_DIR)/dist/files/lib64/wine/vrclient_x64.dll.so /vagrant/vrclient/lib64/wine
+	cp -f $(BUILD_DIR)/dist/files/lib/wine/vrclient.dll.so $(VAGRANT_DIR)/vrclient/lib/wine && \
+	cp -f $(BUILD_DIR)/dist/files/lib64/wine/vrclient_x64.dll.so $(VAGRANT_DIR)/vrclient/lib64/wine
 
 wineopenxr: | vagrant_share/wineopenxr/lib/wine
 wineopenxr: | vagrant_share/wineopenxr/lib64/wine
 wineopenxr: private SHELL := $(VAGRANT_SHELL)
 wineopenxr: downloads
 	$(MAKE) $(MFLAGS) $(MAKEOVERRIDES) -C $(BUILD_DIR)/ $(UNSTRIPPED) $(CCACHE_FLAG) wineopenxr && \
-	cp -f $(BUILD_DIR)/dist/files/lib64/wine/wineopenxr.dll.so /vagrant/wineopenxr/lib64/wine
+	cp -f $(BUILD_DIR)/dist/files/lib64/wine/wineopenxr.dll.so $(VAGRANT_DIR)/wineopenxr/lib64/wine
 
 vagrant_share/%:
 	mkdir -p $@
